Move saToken to GoogleStorage
