+++ src/components/IGVBrowser.js
@@ -6,7 +6,8 @@ import { getUserProjectForWorkspace, parseGsUri } from 'src/components/data/data
 import { centeredSpinner, icon } from 'src/components/icons'
 import IGVAddTrackModal from 'src/components/IGVAddTrackModal'
 import RequesterPaysModal from 'src/components/RequesterPaysModal'
-import { Ajax, saToken } from 'src/libs/ajax'
+import { Ajax } from 'src/libs/ajax'
+import { saToken } from 'src/libs/ajax/GoogleStorage'
 import colors from 'src/libs/colors'
 import { reportError, withErrorReporting } from 'src/libs/error'
 import { useCancellation, useOnMount } from 'src/libs/react-utils'
