+++ src/pages/workspaces/workspace/WorkspaceContainer.js
@@ -10,7 +10,8 @@ import { locationTypes } from 'src/components/region-common'
 import { TabBar } from 'src/components/tabBars'
 import TopBar from 'src/components/TopBar'
 import { updateRecentlyViewedWorkspaces } from 'src/components/workspace-utils'
-import { Ajax, saToken } from 'src/libs/ajax'
+import { Ajax } from 'src/libs/ajax'
+import { saToken } from 'src/libs/ajax/GoogleStorage'
 import { getUser } from 'src/libs/auth'
 import { isTerra } from 'src/libs/brand-utils'
 import colors from 'src/libs/colors'
